Split avg render-tree into one class per file
Split the monolithic src/avg/include/avg/rendertree.h (479 lines) and
src/avg/src/rendertree.cpp (977 lines) into one .h/.cpp pair per class,
matching the project's one-class-per-file layout. Per-class headers live
under avg/rendertree/; the ShapeNode template is header-only.

avg/rendertree.h is now a thin umbrella that includes every per-class
header, so the existing consumers are unchanged.

Pure move: code and comments carried over verbatim, no behavior change.
