+++ src/avg/meson.build
@@ -3,7 +3,15 @@ libavgsrc = [
   'src/animated.cpp',
   'src/lerp.cpp',
   'src/drawcontext.cpp',
-  'src/rendertree.cpp',
+  'src/rendertree/uniqueid.cpp',
+  'src/rendertree/updateresult.cpp',
+  'src/rendertree/rendertreenode.cpp',
+  'src/rendertree/containernode.cpp',
+  'src/rendertree/rectnode.cpp',
+  'src/rendertree/transitionnode.cpp',
+  'src/rendertree/clipnode.cpp',
+  'src/rendertree/idnode.cpp',
+  'src/rendertree/rendertree.cpp',
   'src/fontmanager.cpp',
   'src/fontimpl.cpp',
   'src/font.cpp',
